fixed circular import & serialized page-range
